Implement online service tests

+++ conversion/converter.py
@@ -13,7 +13,7 @@ class Converter:
         # Convert type to float and reduce it values into [0, 255] size range.
         im = im.pixel_array.astype(float)
         im = (np.maximum(im, 0) / np.amax(im)) * 255
-        im = np.uint(im)
+        im = np.uint8(im)
         im = Image.fromarray(im)
         img_byte_arr = io.BytesIO()
         im.save(img_byte_arr, 'PNG')
